refactor(ai): update `ollama`

+++ ollama.sh
@@ -1,16 +1,25 @@
 #!/usr/bin/env bash
 
-# ┏━┓╻  ╻  ┏━┓┏┳┓┏━┓
-# ┃ ┃┃  ┃  ┣━┫┃┃┃┣━┫
-# ┗━┛┗━╸┗━╸╹ ╹╹ ╹╹ ╹
-# start ollama server
+#   ┏━┓╻  ╻  ┏━┓┏┳┓┏━┓
+#   ┃ ┃┃  ┃  ┣━┫┃┃┃┣━┫
+#   ┗━┛┗━╸┗━╸╹ ╹╹ ╹╹ ╹
+# -- start ollama server --
 
-set -euo pipefail
+set -uo pipefail
 
 PROG="${0##*/}"
-DEPS=(podman tmux fzf)
+DEPS=(podman fzf)
 PODIMG=docker.io/ollama/ollama
-TMUX_SESSION="ollama"
+IMAGE="ollama"
+TMUX_SESSION="$IMAGE"
+FZFARGS=(--tmux --no-preview)
+
+# colors
+GREEN=$(tput setaf 2)
+RED=$(tput setaf 1)
+MAGENTA=$(tput setaf 5)
+BOLD=$(tput bold)
+RESET="$(tput sgr0)"
 
 # shellcheck disable=SC2034
 PODCOMPOSE="
@@ -30,21 +39,14 @@ volumes:
 "
 
 if [[ "${1:-}" == "stop" ]]; then
-    podman stop ollama
+    echo -n "$PROG: stopping server..."
+    podman stop ollama >/dev/null
+    echo -e "${GREEN}done${RESET}"
     exit 0
 fi
 
-# colors
-GREEN=$(tput setaf 2)
-RED=$(tput setaf 1)
-MAGENTA=$(tput setaf 5)
-BOLD=$(tput bold)
-RESET="$(tput sgr0)"
-
 function _logerr {
     printf "%s: %b%s%b" "$PROG" "$RED" "$1" "$RESET"
-    echo
-    read -p "Press ENTER to continue..." -r _
     exit 1
 }
 
@@ -64,11 +66,19 @@ function _has {
     done
 }
 
+function _exec {
+    podman exec -it "$IMAGE" ollama "$@"
+}
+
 function _select_model {
-    local model
-    local models=(llama3.2 deepseek-r1)
-    model=$(printf "%s\n" "${models[@]}" | fzf --tmux --no-preview --prompt="model> ")
-    [[ -z "$model" ]] && exit 1
+    local model models
+    models=$(_exec list | awk -F':' '{if (NR > 1) print $1}')
+    FZFARGS+=(--prompt="model> ")
+    model=$(printf "%s\n" "${models[@]}" | fzf "${FZFARGS[@]}")
+    if [[ -z "$model" ]]; then
+        [[ -n "${TMUX:-}" ]] && tmux kill-session -t "$TMUX_SESSION"
+        return
+    fi
 
     echo "$model"
 }
@@ -78,62 +88,58 @@ function _run {
     [[ -z "$model" ]] && exit 1
     printf "ollama: running model %b%b'%s'%b\n\n" "$BOLD" "$MAGENTA" "$model" "$RESET"
     podman exec -it ollama ollama run "$model"
-    tmux kill-session -t "$TMUX_SESSION"
 }
 
 function _start_server {
-    local started
-
-    if [[ -z "$XDG_RUNTIME_DIR" ]]; then
-        _logerr "XDG_RUNTIME_DIR not set"
-    fi
-
-    mkdir -p "$XDG_RUNTIME_DIR/containers/networks/rootless-netns/run" >/dev/null
+    local retcode
     podman start ollama >/dev/null
-    started=$?
-    if [[ "$started" -eq 0 ]]; then
-        printf "%b%s%b\n\n" "$GREEN" "done" "$RESET"
+    retcode=$?
+    if [[ "$retcode" -eq 0 ]]; then
+        printf "%b%s%b\n" "$GREEN" "done" "$RESET"
     else
         _logerr "could not start ollama server"
     fi
 }
 
 function _server_running {
-    if ! podman ps --format "{{.Names}}" | grep -q "ollama"; then
-        opts=(yes no)
-        action=$(printf "%s\n" "${opts[@]}" | fzf --tmux --no-preview --prompt="start ollama server?> ")
-        if [[ ! "$action" =~ ^(y|Y|yes|Yes|"")$ ]]; then
-            exit 1
-        fi
-
-        echo -n "starting server..."
-        _start_server
+    if podman ps --format "{{.Names}}" | grep -q "ollama"; then
+        return
     fi
+
+    local action
+    local opts=(yes no)
+    FZFARGS+=(--prompt="start ollama server?> ")
+    action=$(printf "%s\n" "${opts[@]}" | fzf "${FZFARGS[@]}")
+    if [[ ! "$action" =~ ^(y|Y|yes|Yes|"")$ || "$?" -gt 0 ]]; then
+        [[ -n "${TMUX:-}" ]] && tmux kill-session -t "$TMUX_SESSION"
+        exit 1
+    fi
+
+    echo -n "$PROG: starting server..."
+    _start_server
 }
 
-_has -v "${DEPS[@]}"
+function main {
+    _has -v "${DEPS[@]}"
 
-# check if the image is found
-if ! podman image exists "$PODIMG"; then
-    _logerr "image '$PODIMG' not found"
-fi
+    # check if the image is found
+    if ! podman image exists "$PODIMG"; then
+        _logerr "image '$PODIMG' not found"
+    fi
 
-# create/attach to session
-if ! tmux has-session -t "$TMUX_SESSION" 2>/dev/null; then
-    tmux new-session -d -s "$TMUX_SESSION" "zsh"
-    tmux send-keys -t "$TMUX_SESSION" "$(declare -f _logerr)" C-m
-    tmux send-keys -t "$TMUX_SESSION" "$(declare -f _server_running)" C-m
-    tmux send-keys -t "$TMUX_SESSION" "$(declare -f _start_server)" C-m
-    tmux send-keys -t "$TMUX_SESSION" 'clear' C-m
-    tmux send-keys -t "$TMUX_SESSION" "_server_running" C-m
-    tmux send-keys -t "$TMUX_SESSION" "$(declare -f _select_model)" C-m
-    tmux send-keys -t "$TMUX_SESSION" "$(declare -f _run)" C-m
-    tmux send-keys -t "$TMUX_SESSION" 'clear' C-m
-    # shellcheck disable=2016
-    tmux send-keys -t "$TMUX_SESSION" 'model=$(_select_model)' C-m
-    tmux send-keys -t "$TMUX_SESSION" 'clear' C-m
-    # shellcheck disable=2016
-    tmux send-keys -t "$TMUX_SESSION" '_run "$model"' C-m
-fi
+    _server_running
+    local model
+    model=$(_select_model)
+    [[ -z "$model" ]] && {
+        echo -n "$PROG: stopping server..."
+        podman stop "$IMAGE" >/dev/null
+        echo -e "${GREEN}done${RESET}"
+        exit 1
+    }
+    if [[ -n "${TMUX:-}" ]]; then
+        tmux rename-window -t "ollama" "$model"
+    fi
+    _run "$model"
+}
 
-tmux attach-session -t "$TMUX_SESSION"
+main
